Hacky fix to ware_t good amount overflow when viewing a stop information window set to list in "via (amount)" order.
If a single via destination had more than 8,388,607 units heading to it an arithmetic overflow would occur causing total unit count to be reported incorrectly. This fix prevents packet mergers once logical amount limits are reached hence keeping counting correct at the cost of duplicate listings.

git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@8677 8aca7d54-2c30-db11-9de9-000461428c89

+++ freight_list_sorter.h
@@ -26,7 +26,7 @@ private:
 
 	static bool compare_ware(ware_t const& w1, ware_t const& w2);
 
-	static void add_ware_heading( cbuffer_t &buf, uint32 sum, uint32 max, const ware_t *ware, const char *what_doing );
+	static void add_ware_heading( cbuffer_t &buf, uint64 sum, uint32 max, const ware_t *ware, const char *what_doing );
 };
